requirements.txt
# 核心依赖
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
openai>=1.0.0  # DeepSeek API 兼容 OpenAI 格式

# GUI 支持（Python 内置 tkinter，无需安装）
# Windows 用户如遇到问题可安装：
# pywin32>=306  # Windows 特定功能

# 打包工具（可选）
# pyinstaller>=6.0  # 打包成 exe

# 搜索相关
duckduckgo-search>=3.9.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16

# 图片处理
Pillow>=10.0.0
requests-toolbelt>=1.0.0

# 飞书 API
lark-oapi>=1.0.0

# 定时任务
schedule>=1.2.0
APScheduler>=3.10.0

# 工具库
tqdm>=4.66.0
loguru>=0.7.0
pydantic>=2.0.0
aiohttp>=3.9.0
asyncio>=3.4.3
